+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Fixed a Python 3.8-ism.
+- Added a ``params`` parameter to :func:`feincms3.plugins.external.oembed_json`
+  which allows overriding values sent to the oEmbed provider.
+
 
 `0.94`_ (2021-09-29)
 ~~~~~~~~~~~~~~~~~~~~